Added data attr., removed label <p> tags
Tooltips content is saved as dataset.tooltip inside <button> tag with a class of action-button.

No more tooltips as <p> tags! anywhere in the app. This practise is kept everywhere.

A certain button's tooltip is displayed in a div with a class of 'div-tooltip'. The js  function with the buttons position and places tooltip.